fix(store): Fix integer overflow in get_into/batch_get_into for values > 4GB (#920)
* Fix integer overflow in get_into and batch_get_into for values > 4GB

Problem:
- get_into() and batch_get_into() returned 'int' (32-bit signed)
- This caused integer overflow when handling values larger than ~2GB
- Values > 4GB would wrap around and return incorrect results

Solution:
- Changed return type from 'int' to 'int64_t' for both methods
- Updated batch_get_into to return std::vector<int64_t>
- Updated test assertions to use int64_t

Impact:
- Can now handle values up to ~9.2 exabytes (2^63 - 1)
- Backward compatible for values < 2GB
- Internal implementation already used int64_t, so this aligns public API
- Python bindings automatically handle int64_t -> Python int conversion

Files changed:
- mooncake-store/include/pybind_client.h
- mooncake-store/src/pybind_client.cpp
- mooncake-store/tests/pybind_client_test.cpp

* docs: Update Doxygen comment for batch_get_into return type

Update the documentation comment to accurately reflect that batch_get_into
returns a vector of 64-bit integers, not just 'integers'. This clarifies
that the return type can handle values larger than 2GB/4GB.

+++ mooncake-store/src/pybind_client.cpp
@@ -835,7 +835,7 @@ tl::expected<int64_t, ErrorCode> PyClient::get_into_internal(
     return static_cast<int64_t>(total_size);
 }
 
-int PyClient::get_into(const std::string &key, void *buffer, size_t size) {
+int64_t PyClient::get_into(const std::string &key, void *buffer, size_t size) {
     return to_py_ret(get_into_internal(key, buffer, size));
 }
 
@@ -951,11 +951,11 @@ int PyClient::put_from(const std::string &key, void *buffer, size_t size,
     return to_py_ret(put_from_internal(key, buffer, size, config));
 }
 
-std::vector<int> PyClient::batch_get_into(const std::vector<std::string> &keys,
-                                          const std::vector<void *> &buffers,
-                                          const std::vector<size_t> &sizes) {
+std::vector<int64_t> PyClient::batch_get_into(
+    const std::vector<std::string> &keys, const std::vector<void *> &buffers,
+    const std::vector<size_t> &sizes) {
     auto internal_results = batch_get_into_internal(keys, buffers, sizes);
-    std::vector<int> results;
+    std::vector<int64_t> results;
     results.reserve(internal_results.size());
 
     for (const auto &result : internal_results) {
